server/requirements.txt
aiohttp==3.9.5
aioredis==1.3.1
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
Django==3.2.13
django-cors-headers==3.10.1
djangorestframework==3.13.1
frozenlist==1.4.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
multidict==6.0.5
openai==0.27.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytz==2024.1
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
zope.interface==6.4.post2
